# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clients/upsertClient": {
            "post": {
                "operationId": "55bf9f37abfabfac3fba7c10f90faac9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "description": "Client Code",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "Client Name",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Client Phone",
                                        "type": "string"
                                    },
                                    "menutitle": {
                                        "description": "Menu Title",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "Address to be displayed om Menu",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Client Email",
                                        "type": "string"
                                    },
                                    "whatsapp": {
                                        "description": "Client Whatsapp Ordering Number",
                                        "type": "string"
                                    },
                                    "facebook": {
                                        "description": "Client Facebook",
                                        "type": "string"
                                    },
                                    "instagram": {
                                        "description": "Client Facebook",
                                        "type": "string"
                                    },
                                    "validtill": {
                                        "description": "Valid Till",
                                        "type": "string"
                                    },
                                    "menulogo": {
                                        "description": "Menu logo ",
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Add/update client"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/clients/get/": {
            "get": {
                "summary": "Get Client Data",
                "operationId": "ccbb2f30d3a02c2eb0136111d80a5285",
                "parameters": [
                    {
                        "name": "clientcode",
                        "in": "query",
                        "description": "Client Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-Custom-Header",
                        "in": "header",
                        "description": "Custom header for additional information",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Body"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    }
}